% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qSage-package.R
\docType{package}
\name{qSage-package}
\alias{qSage}
\alias{qSage-package}
\title{qSage: Utility Functions for Single-Cell and Bioinformatics Workflows}
\description{
Provides utility functions and wrappers for single-cell RNA-seq, Seurat workflows, plotting, metadata summarisation, signature handling, and reproducible bioinformatics analyses.
}
\author{
\strong{Maintainer}: Quang Nguyen \email{quang.nguyen.dukoxblue@gmail.com}

Authors:
\itemize{
  \item Quang Nguyen \email{quang.nguyen.dukoxblue@gmail.com}
}

}
\keyword{internal}
